Add `bin/repository edit` for CLI repository editing
Summary:
Ref T4039. This is mostly to deal with that, to prevent the security issues associated with mutable local paths. The next diff will lock them in the web UI.

I also added a confirmation prompt to `bin/repository delete`, which was a little scary without one.

See one comment inline about the `--as` flag. I don't love this, but when I started adding all the stuff we'd need to let this transaction show up as "Administrator" it quickly got pretty big.

Test Plan: Ran `bin/repository edit ...`, saw an edit with a transaction show up on the web UI.

Reviewers: btrahan

Reviewed By: btrahan

CC: aran

Maniphest Tasks: T4039

Differential Revision: https://secure.phabricator.com/D7579

+++ src/applications/repository/management/PhabricatorRepositoryManagementDeleteWorkflow.php
@@ -15,6 +15,10 @@ final class PhabricatorRepositoryManagementDeleteWorkflow
             'help'        => 'Show additional debugging information.',
           ),
           array(
+            'name'        => 'force',
+            'help'        => 'Do not prompt for confirmation.',
+          ),
+          array(
             'name'        => 'repos',
             'wildcard'    => true,
           ),
@@ -30,9 +34,25 @@ final class PhabricatorRepositoryManagementDeleteWorkflow
     }
 
     $console = PhutilConsole::getConsole();
+
+    if (!$args->getArg('force')) {
+      $console->writeOut("%s\n\n", pht('These repositories will be deleted:'));
+
+      foreach ($repos as $repo) {
+        $console->writeOut(
+          "  %s %s\n",
+          'r'.$repo->getCallsign(),
+          $repo->getName());
+      }
+
+      $prompt = pht('Permanently delete these repositories?');
+      if (!$console->confirm($prompt)) {
+        return 1;
+      }
+    }
+
     foreach ($repos as $repo) {
       $console->writeOut("Deleting '%s'...\n", $repo->getCallsign());
-
       $repo->delete();
     }
